fix: fixes api routes and headers
