+++ apps/web/lib/tags/mutateTags.ts
@@ -1,13 +1,17 @@
+import { TTag } from "@formbricks/types/v1/tags";
 import useSWRMutation from "swr/mutation";
 
 export const useCreateTag = (environmentId: string, productId: string) => {
   const { trigger: createTag, isMutating: isCreatingTag } = useSWRMutation(
     `/api/v1/environments/${environmentId}/product/${productId}/tags`,
-    (url, { arg }: { arg: { responseId: string; name: string } }) => {
+    async (url, { arg }: { arg: { name: string } }): Promise<TTag> => {
       return fetch(url, {
         method: "POST",
-        body: JSON.stringify({ name: arg.name, responseId: arg.responseId }),
-      });
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({ name: arg.name }),
+      }).then((res) => res.json());
     }
   );
 
@@ -15,4 +19,38 @@ export const useCreateTag = (environmentId: string, productId: string) => {
     createTag,
     isCreatingTag,
   };
+};
+
+export const useAddTagToResponse = (environmentId: string, surveyId: string, responseId: string) => {
+  const response = useSWRMutation(
+    `/api/v1/environments/${environmentId}/surveys/${surveyId}/responses/${responseId}/tags`,
+
+    async (url, { arg }: { arg: { tagIdToAdd: string } }) => {
+      return fetch(url, {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({ tagId: arg.tagIdToAdd }),
+      }).then((res) => res.json());
+    }
+  );
+
+  return response;
+};
+
+export const removeTagFromResponse = async (
+  environmentId: string,
+  surveyId: string,
+  responseId: string,
+  tagId: string
+) => {
+  const response = await fetch(
+    `/api/v1/environments/${environmentId}/surveys/${surveyId}/responses/${responseId}/tags/${tagId}`,
+    {
+      method: "DELETE",
+    }
+  );
+
+  return response.json();
 };
